Made a Pac-Man
Pac-Man behaves like the original.
The input checks for the pressed key, but also looks if it can change direction; this is important for Cornering. It also checks if it can walk toward the desired cell.
Movement is split into two functions: for x-axis and y-axis. This is so the Cornering can work.
Pac-Man checks if the desired cell is walkable using the WalkableCells Resource.